requirements.txt
Django==2.2.2
django-rest-framework==0.1.0
djangorestframework==3.9.4
djangorestframework-gis==0.14
psycopg2-binary==2.8.2
pytz==2019.1
sqlparse==0.3.0
